Test: 위키, 커뮤니티 관련 컨테이너 삭제 및 관련 페이지 제작을 통한 이동 로직 변경
