+++ src/page/CommunityPage.js
@@ -6,7 +6,7 @@ import { Div, Section } from "../style/LayoutStyle";
 import HeaderItem from "../component/HeaderItem";
 import GameListContainer from "../container/GameListNavContainer";
 import FooterItem from "../component/FooterItem";
-import GameCommunityContainer from "../container/inGamePage/GameCommunityContainer";
+import GameCommunityContainer from "../container/inCommunityPage/CommunityContainer";
 
 const PageWrapper = styled(Div)`
   min-height: 100vh;
