+++ src/page/WikiHistoryContentPage.js
@@ -6,7 +6,7 @@ import { Div, Section } from "../style/LayoutStyle";
 import HeaderItem from "../component/HeaderItem";
 import GameListContainer from "../container/GameListNavContainer";
 import FooterItem from "../component/FooterItem";
-import WikiHistoryContentContainer from "../container/inWikiComponent/WikiHistoryContentContainer";
+import WikiHistoryContentContainer from "../container/inWikiPage/WikiHistoryContentContainer";
 
 const PageWrapper = styled(Div)`
   min-height: 100vh;
