+++ src/page/WikiHistoryPage.js
@@ -6,7 +6,7 @@ import { Div, Section } from "../style/LayoutStyle";
 import HeaderItem from "../component/HeaderItem";
 import GameListContainer from "../container/GameListNavContainer";
 import FooterItem from "../component/FooterItem";
-import WikiHistoryListContainer from "../container/inWikiComponent/WikiHistoryListContainer";
+import WikiHistoryListContainer from "../container/inWikiPage/WikiHistoryListContainer";
 
 const PageWrapper = styled(Div)`
   min-height: 100vh;
